+++ iam-login-service/src/test/java/it/infn/mw/iam/test/service/IamAccountServiceTests.java
@@ -9,6 +9,7 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Matchers.anyObject;
 import static org.mockito.Matchers.anyString;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -21,6 +22,8 @@ import java.util.Optional;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mitre.oauth2.model.OAuth2AccessTokenEntity;
+import org.mitre.oauth2.model.OAuth2RefreshTokenEntity;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
@@ -28,6 +31,7 @@ import org.mockito.runners.MockitoJUnitRunner;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
+import it.infn.mw.iam.core.time.TimeProvider;
 import it.infn.mw.iam.core.user.DefaultIamAccountService;
 import it.infn.mw.iam.core.user.exception.CredentialAlreadyBoundException;
 import it.infn.mw.iam.core.user.exception.InvalidCredentialException;
@@ -39,6 +43,8 @@ import it.infn.mw.iam.persistence.model.IamSshKey;
 import it.infn.mw.iam.persistence.model.IamX509Certificate;
 import it.infn.mw.iam.persistence.repository.IamAccountRepository;
 import it.infn.mw.iam.persistence.repository.IamAuthoritiesRepository;
+import it.infn.mw.iam.persistence.repository.IamOAuthAccessTokenRepository;
+import it.infn.mw.iam.persistence.repository.IamOAuthRefreshTokenRepository;
 
 @RunWith(MockitoJUnitRunner.class)
 public class IamAccountServiceTests extends IamAccountServiceTestSupport {
@@ -55,6 +61,14 @@ public class IamAccountServiceTests extends IamAccountServiceTestSupport {
   @Mock
   private ApplicationEventPublisher eventPublisher;
 
+  @Mock
+  private TimeProvider timeProvider;
+  
+  @Mock
+  private IamOAuthAccessTokenRepository accessTokenRepo;
+  
+  @Mock
+  private IamOAuthRefreshTokenRepository refreshTokenRepo;
 
   @InjectMocks
   private DefaultIamAccountService accountService;
@@ -706,5 +720,20 @@ public class IamAccountServiceTests extends IamAccountServiceTestSupport {
     verify(accountRepo, times(1)).delete(CICCIO_ACCOUNT);
     verify(accountRepo, times(1)).delete(TEST_ACCOUNT);
     verify(eventPublisher, times(2)).publishEvent(anyObject());
+  }
+  
+  @Test
+  public void testTokensAreRemovedWhenAccountIsRemoved() {
+    OAuth2AccessTokenEntity accessToken  = mock(OAuth2AccessTokenEntity.class);
+    OAuth2RefreshTokenEntity refreshToken = mock(OAuth2RefreshTokenEntity.class);
+    
+    when(accessTokenRepo.findValidAccessTokensForUser(Mockito.eq(CICCIO_USERNAME), anyObject()))
+      .thenReturn(Arrays.asList(accessToken));
+    when(refreshTokenRepo.findValidRefreshTokensForUser(Mockito.eq(CICCIO_USERNAME), anyObject()))
+      .thenReturn(Arrays.asList(refreshToken));
+    
+    accountService.deleteAccount(CICCIO_ACCOUNT);
+    verify(accessTokenRepo).delete(Mockito.eq(accessToken));
+    verify(refreshTokenRepo).delete(Mockito.eq(refreshToken));
   }
 }